requirements.txt
# Indic-OCR Project Dependencies
# Multi-Language OCR System for Indian Regional Scripts

# Core Libraries
numpy>=1.21.0
pandas>=1.3.0
pillow>=9.0.0

# Computer Vision & Image Processing
opencv-python>=4.5.0
scikit-image>=0.19.0

# Deep Learning Frameworks
torch>=1.10.0
torchvision>=0.11.0
tensorflow>=2.8.0

# OCR Engines
paddlepaddle>=2.4.0
paddleocr>=2.6.0
transformers>=4.20.0  # For TrOCR

# Machine Learning
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Web Interface
streamlit>=1.15.0
flask>=2.0.0

# Utilities
tqdm>=4.62.0
python-Levenshtein>=0.12.0  # For CER/WER calculation
jiwer>=2.5.0  # Word Error Rate calculation
pyyaml>=6.0

# Image augmentation
albumentations>=1.3.0
imgaug>=0.4.0

# Jupyter Notebook support
jupyter>=1.0.0
ipykernel>=6.0.0
